feat: Add @mui/icons-material dependency and create DefaultCard component
The code changes include adding the @mui/icons-material dependency to the project by updating the package.json file. It also introduces a new component called DefaultCard, which is a reusable card component with customizable styles and children. The DefaultCard component is used to display content within a styled card element.